+++ client/tws_foundation_client/lib/src/tws_foundation_source.dart
@@ -1,6 +1,7 @@
 import 'package:csm_client/csm_client.dart';
 import 'package:tws_foundation_client/src/services/account_service.dart';
 import 'package:tws_foundation_client/src/services/bases/accounts_service_base.dart';
+import 'package:tws_foundation_client/src/services/bases/truck_inventory_service_base.dart';
 import 'package:tws_foundation_client/src/services/carriers_service.dart';
 import 'package:tws_foundation_client/src/services/drivers_externals_service.dart';
 import 'package:tws_foundation_client/src/services/drivers_service.dart';
@@ -13,6 +14,7 @@ import 'package:tws_foundation_client/src/services/solutions_service.dart';
 import 'package:tws_foundation_client/src/services/trailer_service.dart';
 import 'package:tws_foundation_client/src/services/trailers_externals_service.dart';
 import 'package:tws_foundation_client/src/services/trucks_externals_service.dart';
+import 'package:tws_foundation_client/src/services/trucks_inventories_service.dart';
 import 'package:tws_foundation_client/src/services/trucks_service.dart';
 import 'package:tws_foundation_client/src/services/vehicules_models_service.dart';
 import 'package:tws_foundation_client/src/services/yard_log_service.dart';
@@ -55,6 +57,8 @@ final class TWSFoundationSource extends CSMSourceBase {
   late final CarriersServiceBase carriers;
   // Vehicules models service
   late final VehiculesModelsServiceBase vehiculesModels;
+  // Truck Inventory service
+  late final TrucksInventoriesServiceBase trucksInventories;
 
   /// Generates a new data source building its internal
   /// services.
@@ -78,6 +82,7 @@ final class TWSFoundationSource extends CSMSourceBase {
     YardLogsService? yardLogs,
     CarriersService? carriers,
     VehiculesModelsService? vehiculesModels,
+    TrucksInventoriesService? trucksInventories,
     CSMUri development = const CSMUri(
       '127.0.0.1',
       '',
@@ -107,5 +112,7 @@ final class TWSFoundationSource extends CSMSourceBase {
     this.yardLogs = yardLogs ?? YardLogsService(host, client: client);
     this.carriers = carriers ?? CarriersService(host, client: client);
     this.vehiculesModels = vehiculesModels ?? VehiculesModelsService(host, client: client);
+    this.trucksInventories = trucksInventories ?? TrucksInventoriesService(host, client: client);
+
   }
 }