+++ server/TWS Foundation/TWS Business/TWSBusinessDatabase.cs
@@ -78,6 +78,8 @@ public partial class TWSBusinessDatabase : BDatabaseSQLS<TWSBusinessDatabase> {
 
     public virtual DbSet<TrailerType> TrailersTypes { get; set; }
 
+    public virtual DbSet<TrailerType> TrucksInventories { get; set; }
+
 
 
     protected override void OnModelCreating(ModelBuilder builder) {
@@ -115,6 +117,7 @@ public partial class TWSBusinessDatabase : BDatabaseSQLS<TWSBusinessDatabase> {
         UsdotH.CreateModel(builder);
         ApproachesH.CreateModel(builder);
         CarrierH.CreateModel(builder);
+        TruckInventory.CreateModel(builder);
 
 
         OnModelCreatingPartial(builder);
@@ -150,6 +153,7 @@ public partial class TWSBusinessDatabase : BDatabaseSQLS<TWSBusinessDatabase> {
             new YardLog(),
             new VehiculeModel(),
             new TrailerType(),
+            new TruckInventory(),
             new UsdotH(),
             new Approach(),
             new ApproachesH(),